+++ example/basic_example_page.dart
@@ -6,25 +6,50 @@ class BasicExamplePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-        child: Baseplate.row(
-            gutter: 8,
-            runSpacing: 8,
-            children: [
-              Baseplate.col(
-                w360: 4,
-                child: Container(color: Colors.red, height: 100),
+    return Scaffold(
+        appBar: AppBar(
+          title: Text('flutter-baseplate Example'),
+        ),
+        body: Container(
+            child: bpRow(
+          runSpacing: 8,
+          padding: 8,
+          gutter: 8,
+          children: [
+            bpCol(
+              w1200: 6,
+              child: Container(
+                color: Colors.red,
+                height: 100,
               ),
-              Baseplate.col(
-                w360: 4,
-                child: Container(color: Colors.red, height: 100),
-              ),
-              Baseplate.col(
-                w360: 4,
-                child: Container(color: Colors.red, height: 100),
-              ),
-            ]
-        )
-    );
+            ),
+            bpCol(
+                w1200: 6,
+                child: bpRow(
+                  children: [
+                    bpCol(
+                      w1200: 6,
+                      child: Container(
+                        color: Colors.pink,
+                        height: 100,
+                      ),
+                    ),
+                    bpCol(
+                      w1200: 6,
+                      child: Container(
+                        color: Colors.blue,
+                        height: 100,
+                      ),
+                    )
+                  ],
+                )),
+            bpCol(
+                w1200: 6,
+                child: Container(
+                  color: Colors.yellow,
+                  height: 100,
+                ))
+          ],
+        )));
   }
 }