+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.1.2] - 06/02/2019
+
+* Added deprecation notice to Baseplate.row and Baseplate.col methods. Calling via bpRow and bpCol is preferred as these widgets _do_ appear in the inspector.
+* Updated example.
+
+## [0.1.1] - 06/02/2019
+
+* Fixed logical error causing padding to display incorrectly when wrapping onto two rows.
+
 ## [0.1.0] - 06/02/2019
 
 * Switched to FractionallySizedBoxes to enable nesting of rows. Could break layouts on some devices hence the version increase.